Responsable: Nahuel Alberto Cachambi user: CachaSan10 Descripción:Se agrego comentario a la clase club en el paquete model.
